fix: add in users comp. to post,delete requests condition (toggle button, user.followed)

+++ src/components/Users/Users.jsx
@@ -9,7 +9,7 @@ import styles from './Users.module.css';
 
 
 const Users = (props) => {
-
+  
   let pagesCount = Math.ceil(props.totalUsersCount / props.pageSize);
   let pages = [];
 
@@ -28,20 +28,21 @@ const Users = (props) => {
       </div>
       <div>
         <button className={styles.followBtn} onClick={() => {
-          axios.post(`https://social-network.samuraijs.com/api/1.0/follow/${user.id}`, {}, {
-            withCredentials: true,
-            headers: {
-              'API-KEY':'62e00a35-a71b-4a1c-b57a-e377af0a9ee0'
-            }
-          })
-            .then(response => {
-              if (response.data.resultCode === 0) {
-                props.followToggle(user.id)
-              }
-            });
 
-
+          if(user.followed === true){
             axios.delete(`https://social-network.samuraijs.com/api/1.0/follow/${user.id}`, {
+              withCredentials: true,
+              headers: {
+                'API-KEY':'62e00a35-a71b-4a1c-b57a-e377af0a9ee0'
+              }
+            })
+              .then(response => {
+                if (response.data.resultCode === 0) {
+                  props.followToggle(user.id)
+                }
+              });
+          } else {
+            axios.post(`https://social-network.samuraijs.com/api/1.0/follow/${user.id}`, {}, {
             withCredentials: true,
             headers: {
               'API-KEY':'62e00a35-a71b-4a1c-b57a-e377af0a9ee0'
@@ -52,6 +53,8 @@ const Users = (props) => {
                 props.followToggle(user.id)
               }
             });
+
+          }           
         }
         }>
           {user.followed ? 'Unfollow' : 'Follow'}
